Add fluidSolver and derive both InsIM and InsIMEX from it
