+++ tests/fsi_leaflet/fsi_leaflet.cpp
@@ -4,9 +4,10 @@
  */
 #include "fsi.h"
 #include "hyperelasticSolver.h"
+#include "insim.h"
 
-extern template class Fluid::NavierStokes<2>;
-extern template class Fluid::NavierStokes<3>;
+extern template class Fluid::InsIM<2>;
+extern template class Fluid::InsIM<3>;
 extern template class Solid::HyperelasticSolver<2>;
 extern template class Solid::HyperelasticSolver<3>;
 extern template class FSI<2>;
@@ -37,7 +38,7 @@ int main(int argc, char *argv[])
             Point<2>(0, 0),
             Point<2>(L, H),
             true);
-          Fluid::NavierStokes<2> fluid(fluid_tria, params);
+          Fluid::InsIM<2> fluid(fluid_tria, params);
 
           Triangulation<2> solid_tria;
           dealii::GridGenerator::subdivided_hyper_rectangle(
